backend/requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
requests
cors
faiss-cpu 
sentence-transformers